add chuong 7 va 8

+++ chuong_3_bai_nang_cao_1/Source.cpp
@@ -6,7 +6,7 @@
 void readNumber(int INumber)
 {
 	int IValueSide = 0;
-	int IExponents = log10(double(INumber)); // So mu.
+	int IExponents = log10(INumber); // So mu.
 	int IQuantityNumber = IExponents + 1; // So luong chu so.
 
 	if (IQuantityNumber > 9)
